wip: pickle/unpickle
