+++ src/graph/heterograph.cc
@@ -7,14 +7,10 @@
 #include <dmlc/io.h>
 #include <dmlc/type_traits.h>
 #include <dgl/array.h>
-#include <dgl/packed_func_ext.h>
-#include <dgl/runtime/container.h>
 #include <dgl/immutable_graph.h>
 #include <vector>
 #include <tuple>
 #include <utility>
-#include "../c_api_common.h"
-#include "./unit_graph.h"
 #include "graph_serializer.h"
 
 using namespace dgl::runtime;
@@ -314,218 +310,6 @@ FlattenedHeteroGraphPtr HeteroGraph::Flatten(const std::vector<dgl_type_t>& etyp
   return FlattenedHeteroGraphPtr(result);
 }
 
-HeteroSubgraph InEdgeGraph(const HeteroGraphPtr graph, const std::vector<IdArray>& vids) {
-  CHECK_EQ(vids.size(), graph->NumVertexTypes())
-    << "Invalid input: the input list size must be the same as the number of vertex types.";
-  std::vector<HeteroGraphPtr> subrels(graph->NumEdgeTypes());
-  std::vector<IdArray> induced_edges(graph->NumEdgeTypes());
-  for (dgl_type_t etype = 0; etype < graph->NumEdgeTypes(); ++etype) {
-    auto pair = graph->meta_graph()->FindEdge(etype);
-    const dgl_type_t src_vtype = pair.first;
-    const dgl_type_t dst_vtype = pair.second;
-    auto relgraph = graph->GetRelationGraph(etype);
-    if (vids[dst_vtype]->shape[0] == 0) {
-      // create a placeholder graph
-      subrels[etype] = UnitGraph::Empty(
-        relgraph->NumVertexTypes(),
-        graph->NumVertices(src_vtype),
-        graph->NumVertices(dst_vtype),
-        graph->DataType(), graph->Context());
-      induced_edges[etype] = IdArray::Empty({0}, graph->DataType(), graph->Context());
-    } else {
-      const auto& earr = graph->InEdges(etype, {vids[dst_vtype]});
-      subrels[etype] = UnitGraph::CreateFromCOO(
-        relgraph->NumVertexTypes(),
-        graph->NumVertices(src_vtype),
-        graph->NumVertices(dst_vtype),
-        earr.src,
-        earr.dst);
-      induced_edges[etype] = earr.id;
-    }
-  }
-  HeteroSubgraph ret;
-  ret.graph = CreateHeteroGraph(graph->meta_graph(), subrels);
-  ret.induced_edges = std::move(induced_edges);
-  return ret;
-}
-
-HeteroSubgraph OutEdgeGraph(const HeteroGraphPtr graph, const std::vector<IdArray>& vids) {
-  CHECK_EQ(vids.size(), graph->NumVertexTypes())
-    << "Invalid input: the input list size must be the same as the number of vertex types.";
-  std::vector<HeteroGraphPtr> subrels(graph->NumEdgeTypes());
-  std::vector<IdArray> induced_edges(graph->NumEdgeTypes());
-  for (dgl_type_t etype = 0; etype < graph->NumEdgeTypes(); ++etype) {
-    auto pair = graph->meta_graph()->FindEdge(etype);
-    const dgl_type_t src_vtype = pair.first;
-    const dgl_type_t dst_vtype = pair.second;
-    auto relgraph = graph->GetRelationGraph(etype);
-    if (vids[src_vtype]->shape[0] == 0) {
-      // create a placeholder graph
-      subrels[etype] = UnitGraph::Empty(
-        relgraph->NumVertexTypes(),
-        graph->NumVertices(src_vtype),
-        graph->NumVertices(dst_vtype),
-        graph->DataType(), graph->Context());
-      induced_edges[etype] = IdArray::Empty({0}, graph->DataType(), graph->Context());
-    } else {
-      const auto& earr = graph->OutEdges(etype, {vids[src_vtype]});
-      subrels[etype] = UnitGraph::CreateFromCOO(
-          relgraph->NumVertexTypes(),
-          graph->NumVertices(src_vtype),
-          graph->NumVertices(dst_vtype),
-          earr.src,
-          earr.dst);
-      induced_edges[etype] = earr.id;
-    }
-  }
-  HeteroSubgraph ret;
-  ret.graph = CreateHeteroGraph(graph->meta_graph(), subrels);
-  ret.induced_edges = std::move(induced_edges);
-  return ret;
-}
-
-HeteroGraphPtr DisjointUnionHeteroGraph(
-    GraphPtr meta_graph, const std::vector<HeteroGraphPtr>& component_graphs) {
-  CHECK_GT(component_graphs.size(), 0) << "Input graph list is empty";
-  std::vector<HeteroGraphPtr> rel_graphs(meta_graph->NumEdges());
-
-  // Loop over all canonical etypes
-  for (dgl_type_t etype = 0; etype < meta_graph->NumEdges(); ++etype) {
-    auto pair = meta_graph->FindEdge(etype);
-    const dgl_type_t src_vtype = pair.first;
-    const dgl_type_t dst_vtype = pair.second;
-    dgl_id_t src_offset = 0, dst_offset = 0;
-    std::vector<dgl_id_t> result_src, result_dst;
-
-    // Loop over all graphs
-    for (size_t i = 0; i < component_graphs.size(); ++i) {
-      const auto& cg = component_graphs[i];
-      EdgeArray edges = cg->Edges(etype);
-      size_t num_edges = cg->NumEdges(etype);
-      const dgl_id_t* edges_src_data = static_cast<const dgl_id_t*>(edges.src->data);
-      const dgl_id_t* edges_dst_data = static_cast<const dgl_id_t*>(edges.dst->data);
-
-      // Loop over all edges
-      for (size_t j = 0; j < num_edges; ++j) {
-        // TODO(mufei): Should use array operations to implement this.
-        result_src.push_back(edges_src_data[j] + src_offset);
-        result_dst.push_back(edges_dst_data[j] + dst_offset);
-      }
-      // Update offsets
-      src_offset += cg->NumVertices(src_vtype);
-      dst_offset += cg->NumVertices(dst_vtype);
-    }
-    HeteroGraphPtr rgptr = UnitGraph::CreateFromCOO(
-      (src_vtype == dst_vtype)? 1 : 2,
-      src_offset,
-      dst_offset,
-      aten::VecToIdArray(result_src),
-      aten::VecToIdArray(result_dst));
-    rel_graphs[etype] = rgptr;
-  }
-  return HeteroGraphPtr(new HeteroGraph(meta_graph, rel_graphs));
-}
-
-std::vector<HeteroGraphPtr> DisjointPartitionHeteroBySizes(
-    GraphPtr meta_graph, HeteroGraphPtr batched_graph, IdArray vertex_sizes, IdArray edge_sizes) {
-  // Sanity check for vertex sizes
-  const uint64_t len_vertex_sizes = vertex_sizes->shape[0];
-  const uint64_t* vertex_sizes_data = static_cast<uint64_t*>(vertex_sizes->data);
-  const uint64_t num_vertex_types = meta_graph->NumVertices();
-  const uint64_t batch_size = len_vertex_sizes / num_vertex_types;
-  // Map vertex type to the corresponding node cum sum
-  std::vector<std::vector<uint64_t>> vertex_cumsum;
-  vertex_cumsum.resize(num_vertex_types);
-  // Loop over all vertex types
-  for (uint64_t vtype = 0; vtype < num_vertex_types; ++vtype) {
-    vertex_cumsum[vtype].push_back(0);
-    for (uint64_t g = 0; g < batch_size; ++g) {
-      // We've flattened the number of vertices in the batch for all types
-      vertex_cumsum[vtype].push_back(
-        vertex_cumsum[vtype][g] + vertex_sizes_data[vtype * batch_size + g]);
-    }
-    CHECK_EQ(vertex_cumsum[vtype][batch_size], batched_graph->NumVertices(vtype))
-      << "Sum of the given sizes must equal to the number of nodes for type " << vtype;
-  }
-
-  // Sanity check for edge sizes
-  const uint64_t* edge_sizes_data = static_cast<uint64_t*>(edge_sizes->data);
-  const uint64_t num_edge_types = meta_graph->NumEdges();
-  // Map edge type to the corresponding edge cum sum
-  std::vector<std::vector<uint64_t>> edge_cumsum;
-  edge_cumsum.resize(num_edge_types);
-  // Loop over all edge types
-  for (uint64_t etype = 0; etype < num_edge_types; ++etype) {
-    edge_cumsum[etype].push_back(0);
-    for (uint64_t g = 0; g < batch_size; ++g) {
-      // We've flattened the number of edges in the batch for all types
-      edge_cumsum[etype].push_back(
-        edge_cumsum[etype][g] + edge_sizes_data[etype * batch_size + g]);
-    }
-    CHECK_EQ(edge_cumsum[etype][batch_size], batched_graph->NumEdges(etype))
-      << "Sum of the given sizes must equal to the number of edges for type " << etype;
-  }
-
-  // Construct relation graphs for unbatched graphs
-  std::vector<std::vector<HeteroGraphPtr>> rel_graphs;
-  rel_graphs.resize(batch_size);
-  // Loop over all edge types
-  for (uint64_t etype = 0; etype < num_edge_types; ++etype) {
-    auto pair = meta_graph->FindEdge(etype);
-    const dgl_type_t src_vtype = pair.first;
-    const dgl_type_t dst_vtype = pair.second;
-    EdgeArray edges = batched_graph->Edges(etype);
-    const dgl_id_t* edges_src_data = static_cast<const dgl_id_t*>(edges.src->data);
-    const dgl_id_t* edges_dst_data = static_cast<const dgl_id_t*>(edges.dst->data);
-    // Loop over all graphs to be unbatched
-    for (uint64_t g = 0; g < batch_size; ++g) {
-      std::vector<dgl_id_t> result_src, result_dst;
-      // Loop over the chunk of edges for the specified graph and edge type
-      for (uint64_t e = edge_cumsum[etype][g]; e < edge_cumsum[etype][g + 1]; ++e) {
-        // TODO(mufei): Should use array operations to implement this.
-        result_src.push_back(edges_src_data[e] - vertex_cumsum[src_vtype][g]);
-        result_dst.push_back(edges_dst_data[e] - vertex_cumsum[dst_vtype][g]);
-      }
-      HeteroGraphPtr rgptr = UnitGraph::CreateFromCOO(
-        (src_vtype == dst_vtype)? 1 : 2,
-        vertex_sizes_data[src_vtype * batch_size + g],
-        vertex_sizes_data[dst_vtype * batch_size + g],
-        aten::VecToIdArray(result_src),
-        aten::VecToIdArray(result_dst));
-      rel_graphs[g].push_back(rgptr);
-    }
-  }
-
-  std::vector<HeteroGraphPtr> rst;
-  for (uint64_t g = 0; g < batch_size; ++g) {
-    rst.push_back(HeteroGraphPtr(new HeteroGraph(meta_graph, rel_graphs[g])));
-  }
-  return rst;
-}
-
-// creator implementation
-HeteroGraphPtr CreateHeteroGraph(
-    GraphPtr meta_graph, const std::vector<HeteroGraphPtr>& rel_graphs) {
-  return HeteroGraphPtr(new HeteroGraph(meta_graph, rel_graphs));
-}
-
-HeteroGraphPtr CreateFromCOO(
-    int64_t num_vtypes, int64_t num_src, int64_t num_dst,
-    IdArray row, IdArray col, SparseFormat restrict_format) {
-  auto unit_g = UnitGraph::CreateFromCOO(
-      num_vtypes, num_src, num_dst, row, col, restrict_format);
-  return HeteroGraphPtr(new HeteroGraph(unit_g->meta_graph(), {unit_g}));
-}
-
-HeteroGraphPtr CreateFromCSR(
-    int64_t num_vtypes, int64_t num_src, int64_t num_dst,
-    IdArray indptr, IdArray indices, IdArray edge_ids,
-    SparseFormat restrict_format) {
-  auto unit_g = UnitGraph::CreateFromCSR(
-      num_vtypes, num_src, num_dst, indptr, indices, edge_ids, restrict_format);
-  return HeteroGraphPtr(new HeteroGraph(unit_g->meta_graph(), {unit_g}));
-}
-
 constexpr uint64_t kDGLSerialize_HeteroGraph = 0xDD589FBE35224ABF;
 
 bool HeteroGraph::Load(dmlc::Stream* fs) {
@@ -558,442 +342,5 @@ void HeteroGraph::Save(dmlc::Stream* fs) const {
     fs->Write(*rptr);
   }
 }
-
-///////////////////////// C APIs /////////////////////////
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroCreateUnitGraphFromCOO")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    int64_t nvtypes = args[0];
-    int64_t num_src = args[1];
-    int64_t num_dst = args[2];
-    IdArray row = args[3];
-    IdArray col = args[4];
-    SparseFormat restrict_format = ParseSparseFormat(args[5]);
-    auto hgptr = CreateFromCOO(nvtypes, num_src, num_dst, row, col, restrict_format);
-    *rv = HeteroGraphRef(hgptr);
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroCreateUnitGraphFromCSR")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    int64_t nvtypes = args[0];
-    int64_t num_src = args[1];
-    int64_t num_dst = args[2];
-    IdArray indptr = args[3];
-    IdArray indices = args[4];
-    IdArray edge_ids = args[5];
-    SparseFormat restrict_format = ParseSparseFormat(args[6]);
-    auto hgptr = CreateFromCSR(nvtypes, num_src, num_dst, indptr, indices, edge_ids,
-                               restrict_format);
-    *rv = HeteroGraphRef(hgptr);
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroCreateHeteroGraph")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    GraphRef meta_graph = args[0];
-    List<HeteroGraphRef> rel_graphs = args[1];
-    std::vector<HeteroGraphPtr> rel_ptrs;
-    rel_ptrs.reserve(rel_graphs.size());
-    for (const auto& ref : rel_graphs) {
-      rel_ptrs.push_back(ref.sptr());
-    }
-    auto hgptr = CreateHeteroGraph(meta_graph.sptr(), rel_ptrs);
-    *rv = HeteroGraphRef(hgptr);
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroGetMetaGraph")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    *rv = GraphRef(hg->meta_graph());
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroGetRelationGraph")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    dgl_type_t etype = args[1];
-    CHECK_LE(etype, hg->NumEdgeTypes()) << "invalid edge type " << etype;
-    // Test if the heterograph is a unit graph.  If so, return itself.
-    auto bg = std::dynamic_pointer_cast<UnitGraph>(hg.sptr());
-    if (bg != nullptr)
-      *rv = bg;
-    else
-      *rv = HeteroGraphRef(hg->GetRelationGraph(etype));
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroGetFlattenedGraph")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    List<Value> etypes = args[1];
-    std::vector<dgl_id_t> etypes_vec;
-    for (Value val : etypes) {
-      // (gq) have to decompose it into two statements because of a weird MSVC internal error
-      dgl_id_t id = val->data;
-      etypes_vec.push_back(id);
-    }
-
-    *rv = FlattenedHeteroGraphRef(hg->Flatten(etypes_vec));
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroDisjointUnion")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    GraphRef meta_graph = args[0];
-    List<HeteroGraphRef> component_graphs = args[1];
-    std::vector<HeteroGraphPtr> component_ptrs;
-    component_ptrs.reserve(component_graphs.size());
-    for (const auto& component : component_graphs) {
-      component_ptrs.push_back(component.sptr());
-    }
-    auto hgptr = DisjointUnionHeteroGraph(meta_graph.sptr(), component_ptrs);
-    *rv = HeteroGraphRef(hgptr);
-});
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroDisjointPartitionBySizes")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    const IdArray vertex_sizes = args[1];
-    const IdArray edge_sizes = args[2];
-    const auto& ret = DisjointPartitionHeteroBySizes(
-      hg->meta_graph(), hg.sptr(), vertex_sizes, edge_sizes);
-    List<HeteroGraphRef> ret_list;
-    for (HeteroGraphPtr hgptr : ret) {
-      ret_list.push_back(HeteroGraphRef(hgptr));
-    }
-    *rv = ret_list;
-});
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroAddVertices")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    dgl_type_t vtype = args[1];
-    int64_t num = args[2];
-    hg->AddVertices(vtype, num);
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroAddEdge")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    dgl_type_t etype = args[1];
-    dgl_id_t src = args[2];
-    dgl_id_t dst = args[3];
-    hg->AddEdge(etype, src, dst);
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroAddEdges")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    dgl_type_t etype = args[1];
-    IdArray src = args[2];
-    IdArray dst = args[3];
-    hg->AddEdges(etype, src, dst);
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroClear")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    hg->Clear();
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroDataType")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    *rv = hg->DataType();
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroContext")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    *rv = hg->Context();
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroNumBits")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    *rv = hg->NumBits();
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroIsMultigraph")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    *rv = hg->IsMultigraph();
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroIsReadonly")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    *rv = hg->IsReadonly();
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroNumVertices")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    dgl_type_t vtype = args[1];
-    *rv = static_cast<int64_t>(hg->NumVertices(vtype));
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroNumEdges")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    dgl_type_t etype = args[1];
-    *rv = static_cast<int64_t>(hg->NumEdges(etype));
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroHasVertex")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    dgl_type_t vtype = args[1];
-    dgl_id_t vid = args[2];
-    *rv = hg->HasVertex(vtype, vid);
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroHasVertices")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    dgl_type_t vtype = args[1];
-    IdArray vids = args[2];
-    *rv = hg->HasVertices(vtype, vids);
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroHasEdgeBetween")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    dgl_type_t etype = args[1];
-    dgl_id_t src = args[2];
-    dgl_id_t dst = args[3];
-    *rv = hg->HasEdgeBetween(etype, src, dst);
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroHasEdgesBetween")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    dgl_type_t etype = args[1];
-    IdArray src = args[2];
-    IdArray dst = args[3];
-    *rv = hg->HasEdgesBetween(etype, src, dst);
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroPredecessors")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    dgl_type_t etype = args[1];
-    dgl_id_t dst = args[2];
-    *rv = hg->Predecessors(etype, dst);
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroSuccessors")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    dgl_type_t etype = args[1];
-    dgl_id_t src = args[2];
-    *rv = hg->Successors(etype, src);
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroEdgeId")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    dgl_type_t etype = args[1];
-    dgl_id_t src = args[2];
-    dgl_id_t dst = args[3];
-    *rv = hg->EdgeId(etype, src, dst);
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroEdgeIds")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    dgl_type_t etype = args[1];
-    IdArray src = args[2];
-    IdArray dst = args[3];
-    const auto& ret = hg->EdgeIds(etype, src, dst);
-    *rv = ConvertEdgeArrayToPackedFunc(ret);
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroFindEdges")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    dgl_type_t etype = args[1];
-    IdArray eids = args[2];
-    const auto& ret = hg->FindEdges(etype, eids);
-    *rv = ConvertEdgeArrayToPackedFunc(ret);
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroInEdges_1")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    dgl_type_t etype = args[1];
-    dgl_id_t vid = args[2];
-    const auto& ret = hg->InEdges(etype, vid);
-    *rv = ConvertEdgeArrayToPackedFunc(ret);
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroInEdges_2")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    dgl_type_t etype = args[1];
-    IdArray vids = args[2];
-    const auto& ret = hg->InEdges(etype, vids);
-    *rv = ConvertEdgeArrayToPackedFunc(ret);
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroOutEdges_1")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    dgl_type_t etype = args[1];
-    dgl_id_t vid = args[2];
-    const auto& ret = hg->OutEdges(etype, vid);
-    *rv = ConvertEdgeArrayToPackedFunc(ret);
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroOutEdges_2")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    dgl_type_t etype = args[1];
-    IdArray vids = args[2];
-    const auto& ret = hg->OutEdges(etype, vids);
-    *rv = ConvertEdgeArrayToPackedFunc(ret);
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroEdges")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    dgl_type_t etype = args[1];
-    std::string order = args[2];
-    const auto& ret = hg->Edges(etype, order);
-    *rv = ConvertEdgeArrayToPackedFunc(ret);
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroInDegree")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    dgl_type_t etype = args[1];
-    dgl_id_t vid = args[2];
-    *rv = static_cast<int64_t>(hg->InDegree(etype, vid));
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroInDegrees")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    dgl_type_t etype = args[1];
-    IdArray vids = args[2];
-    *rv = hg->InDegrees(etype, vids);
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroOutDegree")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    dgl_type_t etype = args[1];
-    dgl_id_t vid = args[2];
-    *rv = static_cast<int64_t>(hg->OutDegree(etype, vid));
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroOutDegrees")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    dgl_type_t etype = args[1];
-    IdArray vids = args[2];
-    *rv = hg->OutDegrees(etype, vids);
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroGetAdj")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    dgl_type_t etype = args[1];
-    bool transpose = args[2];
-    std::string fmt = args[3];
-    *rv = ConvertNDArrayVectorToPackedFunc(
-        hg->GetAdj(etype, transpose, fmt));
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroVertexSubgraph")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    List<Value> vids = args[1];
-    std::vector<IdArray> vid_vec;
-    vid_vec.reserve(vids.size());
-    for (Value val : vids) {
-      vid_vec.push_back(val->data);
-    }
-    std::shared_ptr<HeteroSubgraph> subg(
-        new HeteroSubgraph(hg->VertexSubgraph(vid_vec)));
-    *rv = HeteroSubgraphRef(subg);
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroEdgeSubgraph")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    List<Value> eids = args[1];
-    bool preserve_nodes = args[2];
-    std::vector<IdArray> eid_vec;
-    eid_vec.reserve(eids.size());
-    for (Value val : eids) {
-      eid_vec.push_back(val->data);
-    }
-    std::shared_ptr<HeteroSubgraph> subg(
-        new HeteroSubgraph(hg->EdgeSubgraph(eid_vec, preserve_nodes)));
-    *rv = HeteroSubgraphRef(subg);
-  });
-
-DGL_REGISTER_GLOBAL("transform._CAPI_DGLInSubgraph")
-.set_body([] (DGLArgs args, DGLRetValue *rv) {
-    HeteroGraphRef hg = args[0];
-    const auto& nodes = ListValueToVector<IdArray>(args[1]);
-    std::shared_ptr<HeteroSubgraph> ret(new HeteroSubgraph);
-    *ret = InEdgeGraph(hg.sptr(), nodes);
-    *rv = HeteroGraphRef(ret);
-  });
-
-DGL_REGISTER_GLOBAL("transform._CAPI_DGLOutSubgraph")
-.set_body([] (DGLArgs args, DGLRetValue *rv) {
-    HeteroGraphRef hg = args[0];
-    const auto& nodes = ListValueToVector<IdArray>(args[1]);
-    std::shared_ptr<HeteroSubgraph> ret(new HeteroSubgraph);
-    *ret = OutEdgeGraph(hg.sptr(), nodes);
-    *rv = HeteroGraphRef(ret);
-  });
-
-// HeteroSubgraph C APIs
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroSubgraphGetGraph")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroSubgraphRef subg = args[0];
-    *rv = HeteroGraphRef(subg->graph);
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroSubgraphGetInducedVertices")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroSubgraphRef subg = args[0];
-    List<Value> induced_verts;
-    for (IdArray arr : subg->induced_vertices) {
-      induced_verts.push_back(Value(MakeValue(arr)));
-    }
-    *rv = induced_verts;
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroSubgraphGetInducedEdges")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroSubgraphRef subg = args[0];
-    List<Value> induced_edges;
-    for (IdArray arr : subg->induced_edges) {
-      induced_edges.push_back(Value(MakeValue(arr)));
-    }
-    *rv = induced_edges;
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroAsNumBits")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    int bits = args[1];
-    HeteroGraphPtr hg_new = UnitGraph::AsNumBits(hg.sptr(), bits);
-    *rv = HeteroGraphRef(hg_new);
-  });
-
-DGL_REGISTER_GLOBAL("heterograph_index._CAPI_DGLHeteroCopyTo")
-.set_body([] (DGLArgs args, DGLRetValue* rv) {
-    HeteroGraphRef hg = args[0];
-    int device_type = args[1];
-    int device_id = args[2];
-    DLContext ctx;
-    ctx.device_type = static_cast<DLDeviceType>(device_type);
-    ctx.device_id = device_id;
-    HeteroGraphPtr hg_new = UnitGraph::CopyTo(hg.sptr(), ctx);
-    *rv = HeteroGraphRef(hg_new);
-  });
 
 }  // namespace dgl
